+++ maven-binding-builder/build.gradle.kts
@@ -1,14 +1,9 @@
 plugins {
     buildsrc.convention.`kotlin-jvm`
-    application
 }
 
 dependencies {
-    implementation("org.jetbrains.kotlin:kotlin-compiler:1.9.23")
+    implementation("org.jetbrains.kotlin:kotlin-compiler")
     implementation(projects.actionBindingGenerator)
     runtimeOnly(projects.githubWorkflowsKt)
-}
-
-application {
-    mainClass.set("io.github.typesafegithub.workflows.mavenbinding.MavenBindingBuildingKt")
 }
